reoganization, creeation src folder, main.sh in root
